+++ DataOrientedDesignCodeExample/DataOrietedDesignTest/ood_no_inheritance_non_cont_memory_cars.h
@@ -27,7 +27,7 @@ static void OOD_NoInheritanceNonContMemoryCar(benchmark::State& inState)
     // Initialize the cars
     for (unsigned int car_index = 0; car_index < num_of_cars; ++car_index)
     {
-        const long long id = RandInt();
+        const uint64_t id = RandInt();
         const auto pos = Vector2D(RandFloat(), RandFloat());
         const auto dir = Vector2D(RandFloat(), RandFloat()).Normalized();
         cars.push_back(make_unique<Car>(id, pos, dir));
